major update

+++ src/controllers/UserController.js
@@ -5,6 +5,8 @@ const db = require('../models');
 const bcrypt = require('bcrypt');
 const { v4: uuidv4 } = require('uuid');
 const { validationResult } = require('express-validator');
+const { resizeImage } = require('../helper/ResizeImage');
+const ActivityLog = require('../helper/ActivityLog');
 const User = db.User;
 const Dept = db.Dept;
 const Role = db.Role;
@@ -12,7 +14,7 @@ const Role = db.Role;
 const GetAllUser = async (req, res) => {
   const users = await User.findAll();
 
-  res.render('users/read', {
+  return res.render('users/read', {
     layout: 'layouts/template',
     title: 'List Users',
     users
@@ -37,7 +39,7 @@ const EditUser = async (req, res) => {
 
     const departments = await Dept.findAll();
 
-    res.render('users/edit', {
+    return res.render('users/edit', {
       layout: 'layouts/template',
       title: 'Edit User',
       user,
@@ -45,7 +47,7 @@ const EditUser = async (req, res) => {
       query: req.query       
     });
   } catch (error) {
-    res.status(500).render('errors/500', { layout: false });
+    return res.status(500).render('errors/500', { layout: false });
   }
 };
 
@@ -89,14 +91,21 @@ const UpdateUser = async (req, res) => {
 
     if (!result) {
       await transaction.rollback();
-      return res.render('users/read', { 
-        layout: 'layouts/template', 
-        errors: [{ msg: 'Failed update user' }] 
-      });        
+      req.flash("status", "failed");    
+      req.flash("msg", "Failed Update User.");      
+      return res.redirect(`/users/edit/${user.id}`);
     }
 
+    ActivityLog({
+      account: req.session?.user?.account,
+      username: req.session?.user?.vname,
+      msgs: `User Update Data for (${result.dataValues.empno}) ${result.dataValues.vname}`
+    });
+
     await transaction.commit();
-    res.redirect(`/users/edit/${user.id}?status=success`);
+    req.flash("status", "success");    
+    req.flash("msg", "User Successfully Updated.");
+    return res.redirect(`/users`);
   } catch (error) {
       if (!transaction.finished) {
         await transaction.rollback();
@@ -117,7 +126,7 @@ const CreateUser = async (req, res) => {
   const departments = await Dept.findAll();
   const roles = await Role.findAll();
 
-  res.render('users/create', {
+  return res.render('users/create', {
     layout: 'layouts/template',
     title: 'Add User',
     departments,
@@ -163,14 +172,21 @@ const StoreUser = async (req, res) => {
 
         if (!result) {
           await transaction.rollback();
-          return res.render('users/create', { 
-            layout: 'layouts/template', 
-            errors: [{ msg: 'Failed create user' }] 
-          });            
+          req.flash("status", "failed");    
+          req.flash("msg", "Create User Failed.");                        
+          return res.redirect('/users/create');
         }
 
+        ActivityLog({
+          account: req.session?.user?.account,
+          username: req.session?.user?.vname,
+          msgs: `Created New User (${result.dataValues.empno}) ${result.dataValues.vname}`
+        });
+
         await transaction.commit();
-        res.redirect('/users/create?status=success')
+        req.flash("status", "success");    
+        req.flash("msg", "User Successfully Created.");              
+        return res.redirect('/users')
     } catch (error) {
       if (!transaction.finished) {
         await transaction.rollback();
@@ -207,14 +223,21 @@ const StoreUser = async (req, res) => {
 
       if (!result) {
         await transaction.rollback();
-        return res.render('users/read', { 
-          layout: 'layouts/template', 
-          errors: [{ msg: 'Failed delete user' }] 
-        });        
+        req.flash("status", "failed");    
+        req.flash("msg", "Deleted User Failed.");                
+        return res.redirect('/users');
       }
 
+      ActivityLog({
+        account: req.session?.user?.account,
+        username: req.session?.user?.vname,
+        msgs: `Delete User (${result.dataValues.empno}) ${result.dataValues.vname}`
+      });
+
       await transaction.commit();
-      res.redirect('/users?status=success');
+      req.flash("status", "success");    
+      req.flash("msg", "User Successfully Deleted.");        
+      return res.redirect('/users');
     } catch (error) {
       if (!transaction.finished) {
         await transaction.rollback();
@@ -272,7 +295,7 @@ const UpdateByUser = async (req, res) => {
   try {
     const id = req.params.id;
     const { vname, email } = req.body;
-    const image = req.file.path;
+    const image = req?.file?.path;
     
     const user = await User.findOne({ where: { id }, lock: transaction.LOCK.UPDATE, transaction});
     if (!user) {
@@ -281,20 +304,33 @@ const UpdateByUser = async (req, res) => {
 
       user.vname = vname;
       user.email = email
+
+    if (image) {
       removeFile(user.image);
       user.image = image;
-      const result = await user.save({ transaction });    
+      await resizeImage(user.image);      
+    }      
+
+    const result = await user.save({ transaction });    
 
     if (!result) {
       await transaction.rollback();
-      return res.render('users/account', { 
-        layout: 'layouts/template', 
-        errors: [{ msg: 'Failed update user' }] 
-      });        
+      removeFile(result.dataValues.image);
+      req.flash("status", "failed");    
+      req.flash("msg", "Failed Update Account."); 
+      res.redirect(`/user/account/edit/${user.id}`)
     }
 
+    ActivityLog({
+      account: req.session?.user?.account,
+      username: req.session?.user?.vname,
+      msgs: `Account Updated (${result.dataValues.empno}) ${result.dataValues.vname}`
+    });
+
     await transaction.commit();
-    res.redirect(`/user/account/edit/${user.id}?status=success`)
+    req.flash("status", "success");    
+    req.flash("msg", "Account Successfully Updated."); 
+    res.redirect(`/user/account/edit/${user.id}`)
   } catch (error) {
       if (!transaction.finished) {
         await transaction.rollback();
@@ -315,7 +351,7 @@ const ChangePassword = async (req, res) => {
   try {
     const id = req.params.id;
     const user = await User.findOne({
-      where: { id}});
+      where: { id }});
 
     if (!user) {
       return res.status(404).render('errors/404', { layout: false });
@@ -352,6 +388,7 @@ const UpdatePasswordByUsr = async (req, res) => {
   }
 
   const { oldPswd, newPswd } = req.body;
+  const transaction = await sequelize.transaction();
 
   try {
     const match = await bcrypt.compare(oldPswd, user.passw);
@@ -366,9 +403,24 @@ const UpdatePasswordByUsr = async (req, res) => {
     }
 
     const hashedNew = await bcrypt.hash(newPswd, 12);
-    await user.update({ passw: hashedNew });
+    const result = await user.update({ passw: hashedNew }, { transaction });
+    if (!result) {
+      await transaction.rollback();
+      req.flash("status", "failed");    
+      req.flash("msg", "Change Password Failed."); 
+      return res.redirect(`/user/account/change-password/${id}`);
+    }
 
-    return res.redirect(`/user/account/change-password/${id}?status=success`);
+    ActivityLog({
+      account: req.session?.user?.account,
+      username: req.session?.user?.vname,
+      msgs: `(${result.dataValues.empno}) ${result.dataValues.vname} has Changed Password`
+    });
+
+    await transaction.commit();
+    req.flash("status", "success");    
+    req.flash("msg", "Password Successfully changed.");    
+    return res.redirect(`/user/account/change-password/${id}`);
   } catch (err) {
     return res.render('users/change-password', {
       layout: 'layouts/template',
@@ -379,160 +431,6 @@ const UpdatePasswordByUsr = async (req, res) => {
     });
   }
 };
-
-
-const GetAllPic = async (req, res) => {
-  const pics = await User.findAll({ where: {role_id: 3 }})
-  const departments = await Dept.findAll();
-
-  res.render('users/pic', {
-    layout: 'layouts/template',
-    title: 'List PIC',
-    departments,
-    pics
-  });
-}
-
-const CreatePic = async (req, res) => {
-  const departments = await Dept.findAll();
-
-  res.render('users/pic_create', {
-    layout: 'layouts/template',
-    title: 'Add PIC',
-    departments,
-    errors: [],
-    query: req.query,
-    oldData: {}
-  });
-};
-
-const StorePic = async (req, res) => {
-    const departments = await Dept.findAll();
-    const transaction = await sequelize.transaction();
-
-    try {
-        const errors = validationResult(req);
-        errors.throw();
-
-        const { vname, empno, storeh, sex, dept_id, code } = req.body
-
-        const manuf = 'J'
-        const account = 'J' + empno
-        const passw = "PASSWORDPIC123"
-        const email = vname.replace(/\s+/g, '') + "@sht.ssbshoes.com"
-        const role_id = 3
-        const saltRounds = 12;
-        const hashedPassw = await bcrypt.hash(passw, saltRounds);        
-
-        const result = await User.create({
-                      id: uuidv4(),
-                      manuf,
-                      account,
-                      empno,
-                      vname,              
-                      passw: hashedPassw,
-                      sex,
-                      dept_id,
-                      email,
-                      role_id,
-                      storeh,
-                      code,
-                    }, { transaction });
-
-        if (!result) {
-          await transaction.rollback();
-          return res.render('users/pic_create', { 
-            layout: 'layouts/template', 
-            errors: [{ msg: 'Failed create pic user' }] 
-          });            
-        }
-
-        await transaction.commit();
-        res.redirect('/pic/create?status=success')
-    } catch (error) {
-      if (!transaction.finished) {
-        await transaction.rollback();
-      }
-
-      let errors = [];
-
-      if (error.errors) {
-        errors = error.errors
-      }
-
-      res.render('users/pic_create', {
-        layout: 'layouts/template',
-        title: 'Add PIC',
-        departments,
-        errors,
-        oldData: req.body
-      });
-    }    
-}
-
-const EditPic = async (req, res) => {
-  try {
-    const id = req.params.id;
-    const user = await User.findOne({
-      where: { id}, 
-      include: [
-        {
-          model: Dept,
-          as: 'dept'
-        }
-      ]});
-
-    if (!user) {
-      return res.status(404).render('errors/404', { layout: false });
-    }
-
-    const departments = await Dept.findAll();
-
-    res.render('users/pic_edit', {
-      layout: 'layouts/template',
-      title: 'Edit PIC',
-      user,
-      departments,
-      query: req.query       
-    });
-  } catch (error) {
-    res.status(500).render('errors/500', { layout: false });
-  }
-};
-
-const RemovePic = async (req, res) => {
-    const transaction = await sequelize.transaction();
-    try {
-      const id = req.params.id;
-      const user = await User.findOne({ where: { id }, lock: transaction.LOCK.UPDATE, transaction });
-
-      if (!user) {
-        return res.status(404).render('errors/404', { layout: false });
-      }
-
-      const result = await user.destroy({ transaction });
-
-      if (!result) {
-        await transaction.rollback();
-        return res.render('users/pic', { 
-          layout: 'layouts/template', 
-          errors: [{ msg: 'Failed delete pic user' }] 
-        });        
-      }
-
-      await transaction.commit();
-      res.redirect('/pic?status=success');
-    } catch (error) {
-      if (!transaction.finished) {
-        await transaction.rollback();
-      }
-
-      return res.render('users/pic', { 
-        layout: 'layouts/template', 
-        errors: [{ msg: error }] 
-      });
-    }
-}
 
 const removeFile = (pathway) => {
   const fileName = path.basename(pathway);
@@ -553,10 +451,5 @@ module.exports = {
   RemoveUser, 
   EditByUser, 
   ChangePassword, 
-  UpdatePasswordByUsr,
-  GetAllPic,
-  CreatePic,
-  StorePic,
-  EditPic,
-  RemovePic
+  UpdatePasswordByUsr
 }
